refactor(brutus): extract banner detection to banners.go
