+++ pkg/brutus/brutus.go
@@ -53,7 +53,6 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"strings"
 	"time"
 
 	"github.com/praetorian-inc/brutus/pkg/badkeys"
@@ -198,85 +197,6 @@ type KeyPlugin interface {
 // which is important for concurrent usage.
 type PluginFactory func() Plugin
 
-// =============================================================================
-// Standard Banners
-// =============================================================================
-
-// standardBanners contains known standard banner patterns for each protocol
-var standardBanners = map[string][]string{
-	"ssh": {
-		"SSH-2.0-OpenSSH",
-		"SSH-2.0-libssh",
-		"SSH-2.0-dropbear",
-	},
-	"telnet": {
-		"Ubuntu",
-		"Debian",
-		"Linux",
-		"FreeBSD",
-	},
-	"ftp": {
-		"220 ProFTPD",
-		"220 (vsFTPd",
-		"220-FileZilla",
-		"220 Pure-FTPd",
-	},
-	"mysql": {
-		"MySQL 5.",
-		"MySQL 8.",
-		"MariaDB 10.",
-		"Percona Server",
-	},
-	"snmp": {
-		"Linux",
-		"Cisco IOS",
-		"Windows",
-		"FreeBSD",
-		"net-snmp",
-		"HP ETHERNET",
-		"APC",
-		"Ubiquiti",
-	},
-}
-
-// IsStandardBanner checks if a banner matches known standard patterns for the protocol.
-// Returns true if the banner is standard (common/default), false if custom/modified.
-//
-// HTTP protocols (http, https, couchdb, elasticsearch, influxdb) always return false
-// to enable LLM analysis, as they have application-specific banners (Grafana, Jenkins,
-// Tomcat) that benefit from LLM credential suggestion.
-//
-// Unknown non-HTTP protocols or empty banners are assumed standard.
-func IsStandardBanner(protocol, banner string) bool {
-	// Empty banner - assume standard
-	if banner == "" {
-		return true
-	}
-
-	// HTTP protocols always need LLM analysis (application-specific banners)
-	if isHTTPProtocol(protocol) {
-		return false
-	}
-
-	// Get patterns for protocol
-	patterns, ok := standardBanners[protocol]
-	if !ok {
-		// Unknown protocol - assume standard
-		return true
-	}
-
-	// Check if banner matches any standard pattern
-	for _, pattern := range patterns {
-		if strings.Contains(banner, pattern) {
-			return true
-		}
-	}
-
-	// No match - custom banner
-	return false
-}
-
-// =============================================================================
 // Configuration Validation
 // =============================================================================
